+++ ersatz/src/main/java/io/github/cjstehno/ersatz/match/HttpMethodMatcher.java
@@ -13,29 +13,50 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.cjstehno.ersatz.impl.matchers;
+package io.github.cjstehno.ersatz.match;
 
 import io.github.cjstehno.ersatz.cfg.HttpMethod;
 import io.github.cjstehno.ersatz.server.ClientRequest;
 import lombok.RequiredArgsConstructor;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
+import org.hamcrest.Matcher;
 
 import static io.github.cjstehno.ersatz.cfg.HttpMethod.ANY;
+import static java.util.Arrays.stream;
+import static org.hamcrest.CoreMatchers.*;
 
 /**
  * A matcher used to match the HTTP method of a request.
  */
-@RequiredArgsConstructor
+@RequiredArgsConstructor(staticName = "methodMatching")
 public class HttpMethodMatcher extends BaseMatcher<ClientRequest> {
 
-    private final HttpMethod method;
+    private final Matcher<HttpMethod> matcher;
+
+    /**
+     * Creates a matcher to match any o the specified request methods.
+     *
+     * @param methods the request methods allowed for a match
+     * @return the method matcher
+     */
+    @SuppressWarnings("unchecked")
+    public static HttpMethodMatcher methodMatching(final HttpMethod... methods) {
+        return methodMatching(
+            anyOf(
+                stream(methods)
+                    .map(m -> m == ANY ? any(HttpMethod.class) : equalTo(m))
+                    .toList().toArray(new Matcher[0])
+            )
+        );
+    }
 
     @Override public boolean matches(final Object actual) {
-        return method == ANY || method == ((ClientRequest) actual).getMethod();
+        return matcher.matches(((ClientRequest) actual).getMethod());
     }
 
     @Override public void describeTo(final Description description) {
-        description.appendText("HTTP method is " + method);
+        description.appendText("HTTP method is ");
+        matcher.describeTo(description);
     }
 }