+++ CHANGES.md
@@ -5,3 +5,4 @@
 * #151: Fixed bug related to query param matching
 * Added predicate function matchers
 * Refactored the matcher system from the API interface down
+* new feature - required global requirements